+++ src/y_doc.rs
@@ -1,0 +1,240 @@
+use pyo3::prelude::*;
+use pyo3::types::PyTuple;
+use yrs::Doc;
+use yrs::OffsetKind;
+use yrs::Options;
+
+use crate::y_array::YArray;
+use crate::y_map::YMap;
+use crate::y_text::YText;
+use crate::y_transaction::YTransaction;
+use crate::y_xml::YXmlElement;
+use crate::y_xml::YXmlText;
+/// A y-py document type. Documents are most important units of collaborative resources management.
+/// All shared collections live within a scope of their corresponding documents. All updates are
+/// generated on per document basis (rather than individual shared type). All operations on shared
+/// collections happen via [YTransaction], which lifetime is also bound to a document.
+///
+/// Document manages so called root types, which are top-level shared types definitions (as opposed
+/// to recursively nested types).
+///
+/// A basic workflow sample:
+///
+/// ```python
+/// from y_py import YDoc
+///
+/// doc = YDoc()
+/// with doc.begin_transaction() as txn:
+///     text = txn.get_text('name')
+///     text.push(txn, 'hello world')
+///     output = text.to_string(txn)
+///     print(output)
+/// ```
+#[pyclass(unsendable)]
+pub struct YDoc(pub Doc);
+
+#[pymethods]
+impl YDoc {
+    /// Doc(id)
+    /// --
+    ///
+    /// Creates a new y-py document. If `id` parameter was passed it will be used as this document
+    /// globally unique identifier (it's up to caller to ensure that requirement). Otherwise it will
+    /// be assigned a randomly generated number.
+    #[new]
+    pub fn new(
+        client_id: Option<u64>,
+        offset_kind: Option<String>,
+        skip_gc: Option<bool>,
+    ) -> PyResult<Self> {
+        let mut options = Options::default();
+        if let Some(client_id) = client_id {
+            options.client_id = client_id;
+        }
+
+        if let Some(raw_offset) = offset_kind {
+            let clean_offset = raw_offset.to_lowercase().replace("-", "");
+            let offset = match clean_offset.as_str() {
+                "utf8" => Ok(OffsetKind::Bytes),
+                "utf16" => Ok(OffsetKind::Utf16),
+                "utf32" => Ok(OffsetKind::Utf32),
+                _ => Err(pyo3::exceptions::PyValueError::new_err(format!(
+                    "'{}' is not a valid offset kind (utf8, utf16, or utf32).",
+                    clean_offset
+                ))),
+            }?;
+            options.offset_kind = offset;
+        }
+
+        if let Some(skip_gc) = skip_gc {
+            options.skip_gc = skip_gc;
+        }
+
+        Ok(YDoc(Doc::with_options(options)))
+    }
+
+    /// Gets globally unique identifier of this `YDoc` instance.
+    #[getter]
+    pub fn id(&self) -> f64 {
+        self.0.client_id as f64
+    }
+
+    /// begin_transaction(/)
+    /// --
+    ///
+    /// Returns a new transaction for this document. y-py shared data types execute their
+    /// operations in a context of a given transaction. Each document can have only one active
+    /// transaction at the time - subsequent attempts will cause exception to be thrown.
+    ///
+    /// Transactions started with `doc.begin_transaction` can be released by deleting the transaction object
+    /// method.
+    ///
+    /// Example:
+    ///
+    /// ```python
+    /// from y_py import YDoc
+    /// doc = YDoc()
+    /// text = doc.get_text('name')
+    /// with doc.begin_transaction() as txn:
+    ///     text.insert(txn, 0, 'hello world')
+    /// ```
+    pub fn begin_transaction(&mut self) -> YTransaction {
+        YTransaction(self.0.transact())
+    }
+
+    pub fn transact(&mut self, callback: PyObject) -> PyResult<PyObject> {
+        let txn = self.begin_transaction();
+        Python::with_gil(|py| {
+            let args = PyTuple::new(py, std::iter::once(txn.into_py(py)));
+            callback.call(py, args, None)
+        })
+    }
+
+    /// Returns a `YMap` shared data type, that's accessible for subsequent accesses using given
+    /// `name`.
+    ///
+    /// If there was no instance with this name before, it will be created and then returned.
+    ///
+    /// If there was an instance with this name, but it was of different type, it will be projected
+    /// onto `YMap` instance.
+    pub fn get_map(&mut self, name: &str) -> YMap {
+        self.begin_transaction().get_map(name)
+    }
+
+    /// Returns a `YXmlElement` shared data type, that's accessible for subsequent accesses using
+    /// given `name`.
+    ///
+    /// If there was no instance with this name before, it will be created and then returned.
+    ///
+    /// If there was an instance with this name, but it was of different type, it will be projected
+    /// onto `YXmlElement` instance.
+    pub fn get_xml_element(&mut self, name: &str) -> YXmlElement {
+        YXmlElement(self.begin_transaction().get_xml_element(name))
+    }
+
+    /// Returns a `YXmlText` shared data type, that's accessible for subsequent accesses using given
+    /// `name`.
+    ///
+    /// If there was no instance with this name before, it will be created and then returned.
+    ///
+    /// If there was an instance with this name, but it was of different type, it will be projected
+    /// onto `YXmlText` instance.
+    pub fn get_xml_text(&mut self, name: &str) -> YXmlText {
+        YXmlText(self.begin_transaction().get_xml_text(name))
+    }
+
+    /// Returns a `YArray` shared data type, that's accessible for subsequent accesses using given
+    /// `name`.
+    ///
+    /// If there was no instance with this name before, it will be created and then returned.
+    ///
+    /// If there was an instance with this name, but it was of different type, it will be projected
+    /// onto `YArray` instance.
+    pub fn get_array(&mut self, name: &str) -> YArray {
+        self.begin_transaction().get_array(name)
+    }
+
+    /// Returns a `YText` shared data type, that's accessible for subsequent accesses using given
+    /// `name`.
+    ///
+    /// If there was no instance with this name before, it will be created and then returned.
+    ///
+    /// If there was an instance with this name, but it was of different type, it will be projected
+    /// onto `YText` instance.
+    pub fn get_text(&mut self, name: &str) -> YText {
+        self.begin_transaction().get_text(name)
+    }
+}
+
+/// Encodes a state vector of a given y-py document into its binary representation using lib0 v1
+/// encoding. State vector is a compact representation of updates performed on a given document and
+/// can be used by `encode_state_as_update` on remote peer to generate a delta update payload to
+/// synchronize changes between peers.
+///
+/// Example:
+///
+/// ```python
+/// from y_py import YDoc, encode_state_vector, encode_state_as_update, apply_update from y_py
+///
+/// # document on machine A
+/// local_doc = YDoc()
+/// local_sv = encode_state_vector(local_doc)
+///
+/// # document on machine B
+/// remote_doc = YDoc()
+/// remote_delta = encode_state_as_update(remote_doc, local_sv)
+///
+/// apply_update(local_doc, remote_delta)
+/// ```
+#[pyfunction]
+pub fn encode_state_vector(doc: &mut YDoc) -> Vec<u8> {
+    doc.begin_transaction().state_vector_v1()
+}
+
+/// Encodes all updates that have happened since a given version `vector` into a compact delta
+/// representation using lib0 v1 encoding. If `vector` parameter has not been provided, generated
+/// delta payload will contain all changes of a current y-py document, working effectively as its
+/// state snapshot.
+///
+/// Example:
+///
+/// ```python
+/// from y_py import YDoc, encode_state_vector, encode_state_as_update, apply_update
+///
+/// # document on machine A
+/// local_doc = YDoc()
+/// local_sv = encode_state_vector(local_doc)
+///
+/// # document on machine B
+/// remote_doc = YDoc()
+/// remote_delta = encode_state_as_update(remote_doc, local_sv)
+///
+/// apply_update(local_doc, remote_delta)
+/// ```
+#[pyfunction]
+pub fn encode_state_as_update(doc: &mut YDoc, vector: Option<Vec<u8>>) -> Vec<u8> {
+    doc.begin_transaction().diff_v1(vector)
+}
+
+/// Applies delta update generated by the remote document replica to a current document. This
+/// method assumes that a payload maintains lib0 v1 encoding format.
+///
+/// Example:
+///
+/// ```python
+/// from y_py import YDoc, encode_state_vector, encode_state_as_update, apply_update
+///
+/// # document on machine A
+/// local_doc = YDoc()
+/// local_sv = encode_state_vector(local_doc)
+///
+/// # document on machine B
+/// remote_doc = YDoc()
+/// remote_delta = encode_state_as_update(remote_doc, local_sv)
+///
+/// apply_update(local_doc, remote_delta)
+/// ```
+#[pyfunction]
+pub fn apply_update(doc: &mut YDoc, diff: Vec<u8>) {
+    doc.begin_transaction().apply_v1(diff);
+}
